Fix: Address compilation errors after weapon system refactor
This commit resolves several compilation errors that arose after the
initial refactoring of the weapon and upgrade systems. The errors
were primarily due to outdated import paths and incorrect field
accesses on component references.

Specific fixes include:

1.  **Unresolved Imports:**
    *   Updated `use` statements in `src/survivor.rs`, `src/items.rs`, `src/game.rs`, and `src/in_game_debug_ui.rs` to correctly import `CircleOfWarding` and `SwarmOfNightmares` from their new locations in `crate::custom_weapons::circle_of_warding` and `crate::custom_weapons::swarm_of_nightmares`.

2.  **Field Access Errors (`E0609`) in `src/items.rs`:**
    *   Applied explicit dereferencing (`(*component).field`) for accessing fields on `Health`, `CircleOfWarding`, and `SwarmOfNightmares` mutable references within the `apply_collected_item_effects_system` function. This resolves ambiguity for the compiler when accessing fields on `&mut Component` types obtained from queries.

3.  **Unused Mutable Variable Warning:**
    *   Removed the unnecessary `mut` keyword from the `new_channeling_comp` variable declaration in `src/survivor.rs` within the `survivor_casting_system` for the manual beam activation case, addressing a compiler warning.

**Note on Verification:**
Similar to the previous changes, persistent environment issues have prevented me from successfully verifying the code. Therefore, full compilation verification of these fixes could not be completed. These changes are provided with the strong assumption that they resolve the targeted errors and the codebase is in a more stable state. Further testing and validation in a stable build environment are highly recommended.
