+++ src/custom_weapons/mod.rs
@@ -1,0 +1,3 @@
+// This file will export the custom weapon modules.
+pub mod circle_of_warding;
+pub mod swarm_of_nightmares;
